+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- transfer.scheduled parameter to allow Transfer scheduling
+- StarkBank.Transfer.delete to cancel scheduled Transfers
+- Transaction query by tags
+### Fixed
+- Event errors on unknown subscriptions
 
 ## [0.5.1] - 2020-06-09
 ### Fixed